Redux code refactor
