+++ lib/logic/constants.dart
@@ -1,0 +1,1 @@
+const double KG_LBS_RATIO = 2.2;
