+++ lib/main.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 import 'package:weight_tracker/logic/actions.dart';
-import 'package:weight_tracker/logic/redux_core.dart';
+import 'package:weight_tracker/logic/middleware.dart';
+import 'package:weight_tracker/logic/reducer.dart';
+import 'package:weight_tracker/logic/redux_state.dart';
 import 'package:weight_tracker/screens/main_page.dart';
 
 void main() {
@@ -10,7 +12,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  final Store store = new Store(stateReducer,
+  final Store store = new Store(reduce,
       initialState: new ReduxState(
           firebaseUser: null,
           mainReference: null,
@@ -19,7 +21,7 @@ class MyApp extends StatelessWidget {
           lastRemovedEntry: null,
           hasEntryBeenRemoved: false,
           unit: 'kg'),
-      middleware: [firebaseMiddleware].toList());
+      middleware: [middleware].toList());
 
   @override
   Widget build(BuildContext context) {
